planner: moved preview for storage yard from common space

+++ src/widget/sidebar/city.cpp
@@ -250,12 +250,12 @@ int widget_sidebar_city_handle_mouse(const mouse* m) {
     data.focus_tooltip_text_id = 0;
     if (city_view_is_sidebar_collapsed()) {
         int x_offset = sidebar_common_get_x_offset_collapsed();
-        handled |= image_buttons_handle_mouse(m, x_offset, 24, button_expand_sidebar, std::size(button_expand_sidebar), &button_id);
+        handled |= image_buttons_handle_mouse(m, x_offset, 24, button_expand_sidebar, (int)std::size(button_expand_sidebar), &button_id);
         if (button_id) {
             data.focus_tooltip_text_id = 12;
         }
 
-        handled |= image_buttons_handle_mouse(m, x_offset, 24, buttons_build_collapsed, std::size(buttons_build_collapsed), &button_id);
+        handled |= image_buttons_handle_mouse(m, x_offset, 24, buttons_build_collapsed, (int)std::size(buttons_build_collapsed), &button_id);
         if (button_id) {
             data.focus_tooltip_text_id = button_id + 19;
         }
@@ -266,17 +266,17 @@ int widget_sidebar_city_handle_mouse(const mouse* m) {
         }
 
         int x_offset = sidebar_common_get_x_offset_expanded();
-        handled |= image_buttons_handle_mouse(m, x_offset, 24, buttons_overlays_collapse_sidebar, std::size(buttons_overlays_collapse_sidebar), &button_id);
+        handled |= image_buttons_handle_mouse(m, x_offset, 24, buttons_overlays_collapse_sidebar, (int)std::size(buttons_overlays_collapse_sidebar), &button_id);
         if (button_id) {
             data.focus_tooltip_text_id = button_id + 9;
         }
 
-        handled |= image_buttons_handle_mouse(m, x_offset, 24, buttons_build_expanded, std::size(buttons_build_expanded), &button_id);
+        handled |= image_buttons_handle_mouse(m, x_offset, 24, buttons_build_expanded, (int)std::size(buttons_build_expanded), &button_id);
         if (button_id) {
             data.focus_tooltip_text_id = button_id + 19;
         }
 
-        handled |= image_buttons_handle_mouse(m, x_offset, 24, buttons_top_expanded, std::size(buttons_top_expanded), &button_id);
+        handled |= image_buttons_handle_mouse(m, x_offset, 24, buttons_top_expanded, (int)std::size(buttons_top_expanded), &button_id);
         if (button_id) {
             data.focus_tooltip_text_id = button_id + 40;
         }
@@ -288,9 +288,9 @@ int widget_sidebar_city_handle_mouse(const mouse* m) {
 
 int widget_sidebar_city_handle_mouse_build_menu(const mouse* m) {
     if (city_view_is_sidebar_collapsed()) {
-        return image_buttons_handle_mouse(m, sidebar_common_get_x_offset_collapsed(), 24, buttons_build_collapsed, std::size(buttons_build_collapsed), 0);
+        return image_buttons_handle_mouse(m, sidebar_common_get_x_offset_collapsed(), 24, buttons_build_collapsed, (int)std::size(buttons_build_collapsed), 0);
     } else {
-        return image_buttons_handle_mouse(m, sidebar_common_get_x_offset_expanded(), 24, buttons_build_expanded, std::size(buttons_build_expanded), 0);
+        return image_buttons_handle_mouse(m, sidebar_common_get_x_offset_expanded(), 24, buttons_build_expanded, (int)std::size(buttons_build_expanded), 0);
     }
 }
